Organized web services in ports and adapters
