--- Registry.Web/Services/IUtils.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Registry.Web.Services
-{
-    public interface IUtils
-    {
-        bool IsOrganizationNameValid(string name);
-    }
-}
