+++ Registry.Web/Services/Ports/IDatasetManager.cs
@@ -1,0 +1,11 @@
+﻿using Registry.Web.Models.DTO;
+
+namespace Registry.Web.Services.Ports
+{
+    public interface IDatasetManager
+    {
+        void CreateDataset(DatasetDto ds);
+        void RemoveDataset(int id);
+
+    }
+}
