+++ Registry.Web/Services/Ports/IUtils.cs
@@ -1,0 +1,7 @@
+﻿namespace Registry.Web.Services.Ports
+{
+    public interface IUtils
+    {
+        bool IsOrganizationNameValid(string name);
+    }
+}
