+++ Registry.Web/Startup.cs
@@ -31,6 +31,8 @@ using Registry.Web.Data;
 using Registry.Web.Data.Models;
 using Registry.Web.Models.DTO;
 using Registry.Web.Services;
+using Registry.Web.Services.Adapters;
+using Registry.Web.Services.Ports;
 
 namespace Registry.Web
 {
